format testing

+++ main/models/mri.py
@@ -4,7 +4,7 @@ from datetime import datetime
 class MRICenter(models.Model):
     name = models.CharField(max_length=200)
     email = models.EmailField()
-    phone = models.IntegerField()
+    phone = models.BigIntegerField()
     address = models.CharField(max_length=200)
     address2 = models.CharField(max_length=200)
     city = models.CharField(max_length=200)
